requirements-ingestion.txt
kafka-python==2.0.2
PyYAML==6.0
requests==2.28.2
